src/main/webapp/app/entities/organisation-service/delete/organisation-service-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { OrganisationServiceService } from '../service/organisation-service.service';

import { OrganisationServiceDeleteDialogComponent } from './organisation-service-delete-dialog.component';

describe('OrganisationService Management Delete Component', () => {
  let comp: OrganisationServiceDeleteDialogComponent;
  let fixture: ComponentFixture<OrganisationServiceDeleteDialogComponent>;
  let service: OrganisationServiceService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [OrganisationServiceDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(OrganisationServiceDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(OrganisationServiceDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(OrganisationServiceService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
